extract composer commands into a library
