+++ lib/composer/composer.dart
@@ -1,0 +1,7 @@
+library composer;
+
+export 'src/install.dart';
+export 'src/require_composer_bin_plugin.dart';
+export 'src/install_psalm.dart';
+export 'src/remove_composer_bin.dart';
+export 'src/version.dart';
